anra-workspace/backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
sqlalchemy
pydantic
